create Reservation Model

+++ database/migrations/2021_05_20_083056_add_column_to_users.php
@@ -14,7 +14,7 @@ class AddColumnToUsers extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('inn_id')->unsigned()->index()->nullable();
+            $table->bigInteger('inn_id')->unsigned()->index()->nullable();
             $table->string('address');
             $table->string('tel');
             $table->date('birthday');
